feat: separate node and web builds and use es6 only
