+++ js/api.js
@@ -6,17 +6,11 @@ const ARG_BOOLEAN = 1 << 4;
 
 const DEFAULT_MODEL = "en-us";
 
-// User can specify a default model, or none at all
+if (typeof Module.modelBase === "undefined") {
+  Module.modelBase = "model/";
+}
 if (typeof Module.defaultModel === "undefined") {
   Module.defaultModel = DEFAULT_MODEL;
-}
-// User can also specify the base URL for models
-if (typeof Module.modelBase === "undefined") {
-  if (ENVIRONMENT_IS_WEB) {
-    Module.modelBase = "model/";
-  } else {
-    Module.modelBase = require("./model/index.js");
-  }
 }
 
 /**
@@ -726,77 +720,6 @@ class Endpointer {
         x > 0 ? x / 0x7fff : x / 0x8000
       );
     } else return null;
-  }
-}
-
-/**
- * Async read some JSON (maybe there is a built-in that does this?)
- */
-async function load_json(path) {
-  if (ENVIRONMENT_IS_WEB) {
-    const response = await fetch(path);
-    if (response.ok) return response.json();
-    else
-      throw new Error("Failed to fetch " + path + " :" + response.statusText);
-  } else {
-    const fs = require("fs/promises");
-    const data = await fs.readFile(path, { encoding: "utf8" });
-    return JSON.parse(data);
-  }
-}
-
-/**
- * Load a file from disk or Internet and make it into an s3file_t.
- */
-async function load_to_s3file(path) {
-  let blob_u8;
-  if (ENVIRONMENT_IS_WEB) {
-    const response = await fetch(path);
-    if (response.ok) {
-      const blob = await response.blob();
-      const blob_buf = await blob.arrayBuffer();
-      blob_u8 = new Uint8Array(blob_buf);
-    } else
-      throw new Error("Failed to fetch " + path + " :" + response.statusText);
-  } else {
-    const fs = require("fs/promises");
-    // FIXME: Should read directly to emscripten memory... how?
-    const blob = await fs.readFile(path);
-    blob_u8 = new Uint8Array(blob.buffer);
-  }
-  const blob_len = blob_u8.length + 1;
-  const blob_addr = Module._malloc(blob_len);
-  if (blob_addr == 0)
-    throw new Error("Failed to allocate " + blob_len + " bytes for " + path);
-  writeArrayToMemory(blob_u8, blob_addr);
-  // Ensure it is NUL-terminated in case someone treats it as a string
-  HEAP8[blob_addr + blob_len] = 0;
-  // But exclude the trailing NUL from file size so it works normally
-  return Module._s3file_init(blob_addr, blob_len - 1);
-}
-
-/**
- * Get a model or model file from the built-in model path.
- *
- * The base path can be set by modifying the `modelBase` property of
- * the module object, at initialization or any other time.  Or you can
- * also just override this function if you have special needs.
- *
- * This function is used by `Decoder` to find the default model, which
- * is equivalent to `Model.modelBase + Model.defaultModel`.
- *
- * @param {string} subpath path to model directory or parameter
- * file, e.g. "en-us", "en-us/variances", etc
- * @returns {string} concatenated path. Note this is a simple string
- * concatenation on the Web, so ensure that `modelBase` has a trailing
- * slash if it is a directory.
- */
-function get_model_path(subpath) {
-  if (ENVIRONMENT_IS_WEB) {
-    return Module.modelBase + subpath;
-  } else {
-    const path = require("path");
-    return path.join(Module.modelBase, subpath);
   }
 }
